pyproject.toml
[project]
name = "ai-exam-qu-rec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
